% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identiconr.R
\name{generateIdenticon}
\alias{generateIdenticon}
\title{Randomly generate identicon pieces for a given seed}
\usage{
generateIdenticon(
  seed = (stats::runif(1, 0, 1e+09)),
  nSquare = 4,
  col = grDevices::rgb(stats::runif(1, 0, 0.95), stats::runif(1, 0, 0.95),
    stats::runif(1, 0, 0.95)),
  back = "white"
)
}
\arguments{
\item{seed}{Either an integer or character seed for generating the identicon. If a character seed then the seed is converted to an integer through a portion of the md5 hash}

\item{nSquare}{An integer giving the dimensions of the identicon to be generated. The identicon will be formed of 'nSquare' x 'nSquare' pieces}

\item{col}{A color for the foreground of the identicon}

\item{back}{A color for the background of the identicon}
}
\value{
A list of three items: 'types' giving a data.frame of the color and shape selected for each position of the identicon, 'centers' giving arbitrary positions for the centers of identicon pieces, 'nSquare' giving the number of squares in the identicon
}
\description{
Take a seed and generate an random pattern identicon.
}
\examples{
id1<-generateIdenticon('emailNumber1@email.com')
id2<-generateIdenticon('emailNumber2@email.com')
id3<-generateIdenticon('emailNumber1@email.com')
all(mapply(identical,id1,id2))
all(mapply(identical,id1,id3))
par(mfrow=c(1,3))
plotIdenticon(id1)
plotIdenticon(id2)
plotIdenticon(id3)

}
