% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identiconr.R
\docType{package}
\name{identiconr}
\alias{identiconr}
\title{Functions to generate identicons}
\description{
Produce a randomly generated pattern identicon for a given seed input
}
\details{
The main functions are:
     \describe{
       \item{\code{\link{generateIdenticon}}:}{to randomly generate an assortment of shapes from a given seed}
       \item{\code{\link{plotIdenticon}}:}{plot the assortment of shapes}
     }
}
\examples{
par(mfrow=c(3,3),mar=c(.1,.1,.1,.1))
for(ii in 1:9)plotIdenticon()
par(mfrow=c(2,2),mar=c(.1,.1,.1,.1))
plotIdenticon(generateIdenticon('test'))
plotIdenticon(generateIdenticon('not test'))
plotIdenticon(generateIdenticon('test'))
plotIdenticon(generateIdenticon('not test'))
}
\author{
Scott Sherrill-Mix, \email{shescott@upenn.edu}
}
