% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identiconr.R
\docType{data}
\name{pieces}
\alias{pieces}
\title{Shapes for indenticon generation}
\format{
A list with each item containing a list of shapes where each item is a list definining the point IDs for a potential shape selected in identicon
}
\source{
system.file("data-raw", "makePieces.R", package = "identiconr")
}
\usage{
pieces
}
\description{
A data.frame containing x, y coordinates and rotation conversions for points used in drawing identicon shapes
}
\keyword{datasets}
