% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identiconr.R
\name{plotIdenticon}
\alias{plotIdenticon}
\title{Plot an identicon for a given assortment of pieces or seed}
\usage{
plotIdenticon(identicon = generateIdenticon(...), ...)
}
\arguments{
\item{identicon}{An identicon list output from generateIdenticon}

\item{...}{Additional arguments to generateIdenticon}
}
\description{
Plot a given identicon pattern
}
\examples{
plotIdenticon(seed='emailNumber1@email.com')
plotIdenticon(seed='emailNumber2@email.com')
plotIdenticon(seed='emailNumber1@email.com')
par(mfrow=c(4,4),mar=c(0,0,0,0))
for(ii in 1:16)plotIdenticon()
}
