% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identiconr.R
\docType{data}
\name{points}
\alias{points}
\title{Coordinates for identicon points}
\format{
A data.frame with columns id, x, y and rot1, rot2,
}
\source{
system.file("data-raw", "makePieces.R", package = "identiconr")
}
\usage{
points
}
\description{
A data.frame containing x, y coordinates and rotation conversions for points used in drawing identicon shapes
}
\keyword{datasets}
